Modularized code and further progress
Put functions that created header, footer, and main content sections in a initialPageLoad.js so that those functiosn that build the html are going to be run first. Started creating homePage.js which is fills main content with content from home page. Next we will focus on having index.js generate and place all dom objects on screen, tab switching, and actually creating the other menu and about/contact tab.