Finished home page and tab switching
Finished styling on the home page and finished the main code for tab switching and loading page content. Created 'pageContentModule' which puts the header, main content container, and footer on the page. It also creates tabBtns and sets up logic for them. Most importantly it acts as a module and stores important page html elements so that we can access later. Created 'loadTabContent', which loads tab content by using information from tab switching buttons. still deciding on window.eventListener, but now we're working on styling the menu section