Finished menuPage
Finished creating the html for menu tab and converted it so that the javascript automatically creates it. Also fixed some css styles. Removed position relative and position absolute stylings, made monospace the font in the body so that all text elements use it, added media query for menu grid and menu cards when screen is at 1100px or lower, added universal transition styling, added padding and font-weight for all buttons, css variable --btn-padding is not only exclusive to buttons but alsothe anchor elements that look like buttons for example the anchor tags in the home-content-cards.